requirements.txt
django
djangorestframework
djangorestframework-simplejwt
elasticsearch-dsl>=7.0.0,<8.0.0